Add files via upload
The prototype version of the BrainBeats log-in page and landing page. Index.html is the first screen that users will see, where they are prompted to log in with their account information or create an account. The button to create an account leads to another web page where users can create an account by filling out all of the required information specified. The files are separated into HTML, CSS, and JS so that it can be updated easily in the future. Furthermore, the style of the page is not set in stone yet and is still set to be modified so that it resembles the Figma prototype of CPSC 490.